requirements.txt
streamlit
tensorflow==2.9.1
keras==2.9.0
keras-vggface==0.6
keras-applications==1.0.8
keras-preprocessing==1.1.2
mtcnn==0.1.0
numpy
opencv-python
pillow
scikit-learn
pyperclip
tqdm
